+++ lib/src/graph.cpp
@@ -1,4 +1,4 @@
-#include "util.hpp"
+#include "graph.hpp"
 
 #include <iostream>
